move src/client-builtins to src/builtins/client-plugin
